Lock down mutations

+++ instructor-code/src/resolvers/Mutation.js
@@ -85,7 +85,17 @@ const Mutation = {
       info
     );
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error('Unable to update the post');
+    }
     return prisma.mutation.updatePost(
       {
         where: {
@@ -96,14 +106,15 @@ const Mutation = {
       info
     );
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
           post: {
@@ -116,7 +127,21 @@ const Mutation = {
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment(
+      {
+        id: args.id,
+        author: {
+          id: userId,
+        },
+      },
+      info
+    );
+
+    if (!commentExists) {
+      throw new Error('Unable to delete comment');
+    }
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -126,7 +151,21 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment(
+      {
+        id: args.id,
+        author: {
+          id: userId,
+        },
+      },
+      info
+    );
+
+    if (!commentExists) {
+      throw new Error('Unable to update comment');
+    }
     return prisma.mutation.updateComment(
       {
         where: {
